moving tests to unit
